Creating TopologicalObject class to capture common behavior between Morse and Morse-Smale complexes, and Contour and Merge Trees. Moving toward new common API and adding testing for ContourTrees.
