+++ topopy/__init__.py
@@ -11,6 +11,7 @@ from __future__ import absolute_import
 # These lines ensure that we do not have to do something like:
 # 'from contrib.MorseSmaleComplex import MorseSmaleComplex' outside of
 # this submodule
+from .TopologicalObject import TopologicalObject
 from .MorseSmaleComplex import MorseSmaleComplex
 from .MergeTree import MergeTree
 from .ContourTree import ContourTree
